requirements.txt
asgiref==3.8.1
Django==5.1.4
django-cors-headers==4.6.0
django-extensions==3.2.3
django-phonenumber-field==8.0.0
django-phonenumbers==1.0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
phonenumbers==8.13.52
pillow==11.0.0
PyJWT==2.10.1
sqlparse==0.5.3
